Increasing cloud offset range
Increasing cloud offset range so that the minimum places the top of the cloud layer at y = -64 and the maximum places the bottom of the cloud layer at y = 320

In other words, it will now be possible to put the cloud layer completely outside the build limits by 64 blocks in either direction. Useful in skyblock, oceanblock, and stoneblock worlds.